+++ RecurrentFF/model/data_scenario/processor.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta, abstractmethod
+
+
+class DataScenarioProcessor(metaclass=ABCMeta):
+    @abstractmethod
+    def brute_force_predict(self):
+        pass
